build: v0.11.0 (BREAKING CHANGE)
Since the underlying Typst is still in pre-v1 state, both its interfaces and semantics may change, and so will this crate. I believe I'm the only one who will use this crate as a dependency, so to adhere to semver, let's change the version number from v1.3.2 to v0.11.0 before I (hopefully) publish this on crates.io. Going forward, the minor number will align with Typst's versioning, and the patch number will keep increasing with changes. Old git tags will be renamed with a prefix of `archive-` to signify their historical meanings.